Update Board.cs
Found error in interpreting a node char to a tile and node position. Fixed by adding LocationSpliterNode function for the first tile that has a node. Began working on a function that finds every tile and node position with OtherTilesWithSameNode.
A few function names added, but not implemented.
Some usefull constants added (readonly).